Adding swagger.yaml and refactoring name for Webapp module
