+++ Webapp/src/main/java/io/swagger/api/impl/MortgagecalculationApiServiceImpl.java
@@ -3,33 +3,25 @@ package io.swagger.api.impl;
 import dev.andrylat.bankingfunctionality.mortgagecalculator.CustomerData;
 import dev.andrylat.bankingfunctionality.mortgagecalculator.MortgagePayment;
 import dev.andrylat.bankingfunctionality.mortgagecalculator.MortgagePaymentImpl;
-import io.swagger.api.*;
-import io.swagger.model.*;
-
+import io.swagger.api.MortgagecalculationApiService;
+import io.swagger.api.NotFoundException;
 import io.swagger.model.InlineResponse2001;
 import io.swagger.model.InlineResponse4001;
 import io.swagger.model.MortgageCalculator;
-
-import java.math.BigDecimal;
-import java.util.Map;
-import java.util.List;
-import io.swagger.api.NotFoundException;
-
-import java.io.InputStream;
-
-import io.swagger.v3.oas.models.security.SecurityScheme;
-import org.glassfish.jersey.media.multipart.FormDataContentDisposition;
+import io.swagger.model.PaymentPeriod;
 
 import javax.ws.rs.core.Response;
 import javax.ws.rs.core.SecurityContext;
-import javax.validation.constraints.*;
-@javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.JavaJerseyServerCodegen", date = "2021-01-25T01:46:26.939Z[GMT]")public class MortgagecalculationApiServiceImpl extends MortgagecalculationApiService {
+import java.math.BigDecimal;
+
+@javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.JavaJerseyServerCodegen", date = "2021-01-25T01:46:26.939Z[GMT]")
+public class MortgagecalculationApiServiceImpl extends MortgagecalculationApiService {
     @Override
     public Response mortgagecalculationPost(MortgageCalculator body, SecurityContext securityContext) throws NotFoundException {
 
         MortgagePayment mortgagePayment = new MortgagePaymentImpl();
-        double paymentPeriodId ;
-           System.err.println("Body = " + body);
+        double paymentPeriodId;
+        System.err.println("Body = " + body);
         try {
             PaymentPeriod paymentPeriod = body.getPayementPeriod();
 
@@ -42,7 +34,7 @@ import javax.validation.constraints.*;
             } else {
                 throw new Exception("Your data have an incorrect format");
             }
-        }catch (Exception ex){
+        } catch (Exception ex) {
             System.err.println("ex" + ex);
             InlineResponse4001 response = new InlineResponse4001();
             response.setValidation(false);
@@ -55,7 +47,7 @@ import javax.validation.constraints.*;
 
         mortgagePayment.setMortgageDetails(customerData);
 
-         mortgagePayment.paymentCalculation();
+        mortgagePayment.paymentCalculation();
 
         InlineResponse2001 response = new InlineResponse2001();
         response.setValidation(true);
