+++ pom.xml
@@ -7,7 +7,7 @@
     <modules>
         <module>Dialogs</module>
         <module>Banking_functionality</module>
-        <module>Swagger_doc</module>
+        <module>Webapp</module>
     </modules>
 
     <groupId>dev.andrylat</groupId>
